Move developer utilities to a new `_develop.py` module and refactor imports.
